+++ lib/components/document/src/Cell.ts
@@ -228,9 +228,7 @@ export class Cell extends Component<CellProps, CellChild> {
 						then ./${QNS.w}tcPr/${QNS.w}gridSpan/@${QNS.w}val/number()
 						else 1,
 					"rowSpan": $rowEnd - $rowStart,
-					"children": array{
-						./(${QNS.w}p)
-					},
+					"children": array{ ./(${QNS.w}p) },
 					"verticalAlignment": ./${QNS.w}tcPr/${QNS.w}vAlign/@${QNS.w}val/string(),
 					"insertion": ./${QNS.w}tcPr/${QNS.w}cellIns/map {
 						"id": @${QNS.w}id/number(),
@@ -238,9 +236,9 @@ export class Cell extends Component<CellProps, CellChild> {
 						"date": @${QNS.w}date/string()
 					},
 					"deletion": ./${QNS.w}tcPr/${QNS.w}cellDel/map {
-							"id": @${QNS.w}id/number(),
-							"author": @${QNS.w}author/string(),
-							"date": @${QNS.w}date/string()
+						"id": @${QNS.w}id/number(),
+						"author": @${QNS.w}author/string(),
+						"date": @${QNS.w}date/string()
 					}
 				}
 			`,
